day_typeを移動
