+++ lib/repository/default_timetable_repository.dart
@@ -1,4 +1,5 @@
 import 'package:buss_timetable/domain/timetable_repository.dart';
+import 'package:buss_timetable/model/day_type.dart';
 import 'package:buss_timetable/model/station_name.dart';
 import 'package:buss_timetable/model/timetable.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
