+++ lib/timetable/timetable_view_model.dart
@@ -1,3 +1,4 @@
+import 'package:buss_timetable/model/day_type.dart';
 import 'package:buss_timetable/model/station_name.dart';
 import 'package:buss_timetable/model/timetable.dart';
 import 'package:buss_timetable/timetable/timetable_ui_state.dart';
